Refactored recording into its own package.
